+++ api/v1/api.go
@@ -3,7 +3,6 @@ package v1
 import (
 	"net/http"
 	"fmt"
-	"errors"
 	"strconv"
 	"encoding/json"
 
@@ -11,36 +10,10 @@ import (
 
 	db "elibot-apiserver/dbproxy"
 	Log "elibot-apiserver/log"
-	"elibot-apiserver/mcserver"
 )
 
 func hello(w http.ResponseWriter, r *http.Request) {
 	WriteSuccessResponse(w, "Welcome to elibot\n")
-}
-
-func test(w http.ResponseWriter, r *http.Request) {
-	WriteSuccessResponse(w, "echo\n")
-}
-
-func testSocket(w http.ResponseWriter, r *http.Request) {
-	var mcs *mcserver.MCserver
-	if mcs = mcserver.GetMcServer(); mcs == nil {
-		WriteInternalServerErrorResponse(w, errors.New("mcserver is not available right now"))
-		return
-	}
-	cmd := "testGo 0 1\n"
-	from := "restapi:testsocket"
-	resp := make(chan mcserver.Response)
-	go mcs.Exec(cmd, from, resp)
-	rr := <-resp
-	res := rr.Result
-	err := rr.Err
-	if err!=nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		fmt.Fprintf(w, err.Error())
-	}
-	
-	WriteSuccessResponse(w, res)
 }
 
 func getAllArc(w http.ResponseWriter, r *http.Request) {
@@ -223,80 +196,4 @@ func getAllZeroPoints(w http.ResponseWriter, r *http.Request) {
 	}
 
 	WriteSuccessResponse(w, res)
-}
-
-func DBBackup(w http.ResponseWriter, r *http.Request) {
-	Log.Debug("starting DB Backup")
-	err := db.DBBackup()
-	if err!=nil {
-		WriteInternalServerErrorResponse(w, err)
-		return
-	}
-
-	WriteSuccessResponse(w, "succeed in backup\n")
-}
-
-func DBList(w http.ResponseWriter, r *http.Request) {
-	Log.Debug("starting DB List")
-	files, err := db.DBList()
-	if err!=nil {
-		WriteInternalServerErrorResponse(w, err)
-		return
-	}
-
-	res, _ := json.Marshal(files)
-	WriteSuccessResponse(w, string(res))
-}
-
-func DBDel(w http.ResponseWriter, r *http.Request) {
-	Log.Debug("starting DB Deletion")
-	vars := mux.Vars(r)
-	err := db.DBDel(vars["name"])
-	if err!=nil {
-		WriteInternalServerErrorResponse(w, err)
-		return
-	}
-
-	WriteSuccessResponse(w, "db detelted\n")
-}
-
-func DBRestore(w http.ResponseWriter, r *http.Request) {
-	Log.Debug("starting DB Restore")
-	vars := mux.Vars(r)
-	err := db.DBRestore(vars["name"])
-	if err!=nil {
-		WriteInternalServerErrorResponse(w, err)
-		return
-	}
-
-	WriteSuccessResponse(w, "succeed in restore\n")
-}
-
-func RegisterV1(r *mux.Router) http.Handler {
-	r.HandleFunc("/", hello).Methods("GET")
-	r.HandleFunc("/v1/test", test).Methods("GET")
-	r.HandleFunc("/v1/testSocket", testSocket).Methods("GET")
-
-	r.HandleFunc("/v1/arc", getAllArc).Methods("GET")
-	r.HandleFunc("/v1/arcparams", getArcParams).Methods("GET").Queries("file_no", "{file_no}", "group", "{group}")
-	r.HandleFunc("/v1/bookprograms", getAllBookprograms).Methods("GET")
-	r.HandleFunc("/v1/enum", getAllEnum).Methods("GET")
-	r.HandleFunc("/v1/extaxis", getAllExtaxis).Methods("GET")
-	r.HandleFunc("/v1/interference", getAllInterference).Methods("GET")
-	//r.HandleFunc("/v1/io", getAllIO).Methods("GET")
-	r.HandleFunc("/v1/metadata", getAllMetadata).Methods("GET").Queries("lang", "{lang}")
-	r.HandleFunc("/v1/params", getParams).Methods("GET")
-	r.HandleFunc("/v1/parameterbyid", getParameterById).Methods("GET").Queries("md_id", "{md_id}")
-	r.HandleFunc("/v1/parameterbygroup", parameterbygroup).Methods("GET").Queries("group", "{group}")
-	r.HandleFunc("/v1/ref", getAllRef).Methods("GET")
-	r.HandleFunc("/v1/toolframe", getAllToolframe).Methods("GET")
-	r.HandleFunc("/v1/userframe", getAllUserframe).Methods("GET")
-	r.HandleFunc("/v1/zeropoints", getAllZeroPoints).Methods("GET")
-
-	/*For DB backup, restore and upgrade*/
-	r.HandleFunc("/v1/db/backup", DBBackup).Methods("POST")
-	r.HandleFunc("/v1/db/backup", DBList).Methods("GET")
-	r.HandleFunc("/v1/db/backup", DBDel).Methods("DELETE").Queries("name", "{name}")
-	r.HandleFunc("/v1/db/restore", DBRestore).Methods("POST").Queries("name", "{name}")
-	return r
 }
